+++ src/alisp/src/definitions/alisp_file_functions.cpp
@@ -20,7 +20,7 @@ namespace alisp
 ALObjectPtr Ffile_open(const ALObjectPtr &t_obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_min_size<1>(t_obj));
-    auto name = eval_check(eval, t_obj, 0, &assert_string<int>);
+    auto name = eval_check(eval, t_obj, 0, &assert_string<size_t>);
 
 
     auto output = contains(t_obj, ":out") ? Qt : Qnil;
@@ -35,7 +35,7 @@ ALObjectPtr Ffile_open(const ALObjectPtr &t_obj, env::Environment *, eval::Evalu
 ALObjectPtr Ffile_close(const ALObjectPtr &t_obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<1>(t_obj));
-    auto file = eval_check(eval, t_obj, 0, &assert_file<int>);
+    auto file = eval_check(eval, t_obj, 0, &assert_file<size_t>);
 
     FileHelpers::close_file(file);
     return Qt;
@@ -46,7 +46,7 @@ ALObjectPtr Ffile_read_line(const ALObjectPtr &t_obj, env::Environment *, eval::
 {
 
     AL_CHECK(assert_size<1>(t_obj));
-    auto file = eval_check(eval, t_obj, 0, &assert_file<int>);
+    auto file = eval_check(eval, t_obj, 0, &assert_file<size_t>);
 
     auto &file_obj = FileHelpers::get_file(file);
 
@@ -65,8 +65,8 @@ ALObjectPtr Ffile_write_line(const ALObjectPtr &t_obj, env::Environment *, eval:
 {
 
     AL_CHECK(assert_size<2>(t_obj));
-    auto file = eval_check(eval, t_obj, 0, &assert_file<int>);
-    auto line = eval_check(eval, t_obj, 1, &assert_string<int>);
+    auto file = eval_check(eval, t_obj, 0, &assert_file<size_t>);
+    auto line = eval_check(eval, t_obj, 1, &assert_string<size_t>);
 
     auto &file_obj = FileHelpers::get_file(file);
 
@@ -85,7 +85,7 @@ ALObjectPtr Ffile_has_more(const ALObjectPtr &t_obj, env::Environment *, eval::E
 
     AL_CHECK(assert_size<1>(t_obj));
 
-    auto file = eval_check(eval, t_obj, 0, &assert_file<int>);
+    auto file = eval_check(eval, t_obj, 0, &assert_file<size_t>);
 
     auto &file_obj = FileHelpers::get_file(file);
