+++ src/alisp/src/definitions/alisp_language.cpp
@@ -139,7 +139,7 @@ static ALObjectPtr handle_exception(EXC p_exc, ALObjectPtr t_obj, env::Environme
             if (t_obj->i(0) != Qnil)
             {
                 env::detail::ScopePushPop cpp{ *env };
-                env->put(t_obj->i(0), make_object(make_int(static_cast<int>(p_exc.tag())), make_string(p_exc.what())));
+                env->put(t_obj->i(0), make_object(make_int(static_cast<size_t>(p_exc.tag())), make_string(p_exc.what())));
             }
             return eval_list(eval, handler, 1);
         }
@@ -214,7 +214,7 @@ ALObjectPtr Fimport(const ALObjectPtr &obj, env::Environment *env, eval::Evaluat
     namespace fs = std::filesystem;
     AL_CHECK(assert_min_size<1>(obj));
 
-    auto mod_sym = eval_check(eval, obj, 0, &assert_symbol<int>);
+    auto mod_sym = eval_check(eval, obj, 0, &assert_symbol<size_t>);
 
     const auto module_name = mod_sym->to_string();
