+++ src/alisp/src/definitions/alisp_lists.cpp
@@ -119,7 +119,7 @@ ALObjectPtr Fcons(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *e
 ALObjectPtr Fcdr(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<1>(obj));
-    auto list = eval_check(eval, obj, 0, &assert_list<int>);
+    auto list = eval_check(eval, obj, 0, &assert_list<size_t>);
 
     return splice(list, 1);
 }
@@ -137,7 +137,7 @@ ALObjectPtr Ftail(const ALObjectPtr &obj, env::Environment *env, eval::Evaluator
 ALObjectPtr Flast(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<1>(obj));
-    auto list = eval_check(eval, obj, 0, &assert_list<int>);
+    auto list = eval_check(eval, obj, 0, &assert_list<size_t>);
 
     return list->i(list->length() - 1);
 }
@@ -145,7 +145,7 @@ ALObjectPtr Flast(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *e
 ALObjectPtr Finit(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<1>(obj));
-    auto list = eval_check(eval, obj, 0, &assert_list<int>);
+    auto list = eval_check(eval, obj, 0, &assert_list<size_t>);
 
     return splice(list, 0, static_cast<ALObject::list_type::difference_type>(list->length() - 1));
 }
@@ -153,7 +153,7 @@ ALObjectPtr Finit(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *e
 ALObjectPtr Fpush(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<2>(obj));
-    auto list = eval_check(eval, obj, 0, &assert_list<int>);
+    auto list = eval_check(eval, obj, 0, &assert_list<size_t>);
 
     auto element = eval->eval(obj->i(1));
 
@@ -165,7 +165,7 @@ ALObjectPtr Fpush(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *e
 ALObjectPtr Fshove(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<2>(obj));
-    auto list = eval_check(eval, obj, 0, &assert_list<int>);
+    auto list = eval_check(eval, obj, 0, &assert_list<size_t>);
 
     auto element = eval->eval(obj->i(1));
 
@@ -188,7 +188,7 @@ ALObjectPtr Fnth(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *ev
     AL_CHECK(assert_size<2>(obj));
     auto list = eval->eval(obj->i(0));
 
-    auto index = eval_check(eval, obj, 1, &assert_int<int>);
+    auto index = eval_check(eval, obj, 1, &assert_int<size_t>);
 
 
     if (pstring(list))
@@ -209,7 +209,7 @@ ALObjectPtr Fnth(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *ev
 ALObjectPtr Ffind(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<2>(obj));
-    auto list = eval_check(eval, obj, 0, &assert_list<int>);
+    auto list = eval_check(eval, obj, 0, &assert_list<size_t>);
 
     auto element = eval->eval(obj->i(1));
 
@@ -227,8 +227,8 @@ ALObjectPtr Ffind(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *e
 ALObjectPtr Finsert(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<3>(obj));
-    auto list  = eval_check(eval, obj, 0, &assert_list<int>);
-    auto index = eval_check(eval, obj, 1, &assert_int<int>);
+    auto list  = eval_check(eval, obj, 0, &assert_list<size_t>);
+    auto index = eval_check(eval, obj, 1, &assert_int<size_t>);
 
     auto el = eval->eval(obj->i(2));
 
@@ -241,7 +241,7 @@ ALObjectPtr Finsert(const ALObjectPtr &obj, env::Environment *, eval::Evaluator 
 ALObjectPtr Fcontains(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<2>(obj));
-    auto list = eval_check(eval, obj, 0, &assert_list<int>);
+    auto list = eval_check(eval, obj, 0, &assert_list<size_t>);
 
     auto element = eval->eval(obj->i(1));
 
@@ -260,7 +260,7 @@ ALObjectPtr Fcontains(const ALObjectPtr &obj, env::Environment *, eval::Evaluato
 ALObjectPtr Fclear(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<1>(obj));
-    auto list = eval_check(eval, obj, 0, &assert_list<int>);
+    auto list = eval_check(eval, obj, 0, &assert_list<size_t>);
 
     list->children().clear();
     return list;
@@ -281,7 +281,7 @@ ALObjectPtr Flist(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *e
 ALObjectPtr Fdelete(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<2>(obj));
-    auto list = eval_check(eval, obj, 0, &assert_list<int>);
+    auto list = eval_check(eval, obj, 0, &assert_list<size_t>);
 
     auto element = eval->eval(obj->i(1));
 
@@ -297,7 +297,7 @@ ALObjectPtr Fdelete(const ALObjectPtr &obj, env::Environment *, eval::Evaluator 
 ALObjectPtr Fremove(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<2>(obj));
-    auto list = eval_check(eval, obj, 0, &assert_list<int>);
+    auto list = eval_check(eval, obj, 0, &assert_list<size_t>);
 
     auto element   = eval->eval(obj->i(1));
     auto &children = list->children();
@@ -316,7 +316,7 @@ ALObjectPtr Fremove(const ALObjectPtr &obj, env::Environment *, eval::Evaluator 
 ALObjectPtr Fdelq(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<2>(obj));
-    auto list = eval_check(eval, obj, 0, &assert_list<int>);
+    auto list = eval_check(eval, obj, 0, &assert_list<size_t>);
 
     auto element = eval->eval(obj->i(1));
 
@@ -332,7 +332,7 @@ ALObjectPtr Fdelq(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *e
 ALObjectPtr Fremq(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<2>(obj));
-    auto list = eval_check(eval, obj, 0, &assert_list<int>);
+    auto list = eval_check(eval, obj, 0, &assert_list<size_t>);
 
     auto element   = eval->eval(obj->i(1));
     auto &children = list->children();
@@ -351,9 +351,9 @@ ALObjectPtr Frange(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *
 {
     AL_CHECK(assert_min_size<2>(obj));
 
-    auto start = eval_check(eval, obj, 0, &assert_int<int>);
+    auto start = eval_check(eval, obj, 0, &assert_int<size_t>);
 
-    auto end = eval_check(eval, obj, 1, &assert_int<int>);
+    auto end = eval_check(eval, obj, 1, &assert_int<size_t>);
 
 
     const auto step = [&obj, &eval]() {
