+++ src/alisp/src/definitions/alisp_math.cpp
@@ -125,9 +125,9 @@ ALObjectPtr Flt(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *eva
 {
     AL_CHECK(assert_size<2>(obj));
 
-    const auto one = eval_check(eval, obj, 0, &assert_number<int>);
+    const auto one = eval_check(eval, obj, 0, &assert_number<size_t>);
 
-    const auto two = eval_check(eval, obj, 1, &assert_number<int>);
+    const auto two = eval_check(eval, obj, 1, &assert_number<size_t>);
 
 
     if (one->to_real() < two->to_real())
@@ -145,9 +145,9 @@ ALObjectPtr Fleq(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *ev
 {
     AL_CHECK(assert_min_size<0>(obj));
 
-    const auto one = eval_check(eval, obj, 0, &assert_number<int>);
+    const auto one = eval_check(eval, obj, 0, &assert_number<size_t>);
 
-    const auto two = eval_check(eval, obj, 1, &assert_number<int>);
+    const auto two = eval_check(eval, obj, 1, &assert_number<size_t>);
 
 
     if (one->to_real() <= two->to_real())
@@ -165,9 +165,9 @@ ALObjectPtr Fgt(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *eva
 {
     AL_CHECK(assert_min_size<0>(obj));
 
-    const auto one = eval_check(eval, obj, 0, &assert_number<int>);
+    const auto one = eval_check(eval, obj, 0, &assert_number<size_t>);
 
-    const auto two = eval_check(eval, obj, 1, &assert_number<int>);
+    const auto two = eval_check(eval, obj, 1, &assert_number<size_t>);
 
 
     if (one->to_real() > two->to_real())
@@ -185,9 +185,9 @@ ALObjectPtr Fgeq(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *ev
 {
     AL_CHECK(assert_min_size<0>(obj));
 
-    const auto one = eval_check(eval, obj, 0, &assert_number<int>);
+    const auto one = eval_check(eval, obj, 0, &assert_number<size_t>);
 
-    const auto two = eval_check(eval, obj, 1, &assert_number<int>);
+    const auto two = eval_check(eval, obj, 1, &assert_number<size_t>);
 
 
     if (one->to_real() >= two->to_real())
@@ -205,9 +205,9 @@ ALObjectPtr Feq_math(const ALObjectPtr &obj, env::Environment *, eval::Evaluator
 {
     AL_CHECK(assert_min_size<0>(obj));
 
-    const auto one = eval_check(eval, obj, 0, &assert_number<int>);
+    const auto one = eval_check(eval, obj, 0, &assert_number<size_t>);
 
-    const auto two = eval_check(eval, obj, 1, &assert_number<int>);
+    const auto two = eval_check(eval, obj, 1, &assert_number<size_t>);
 
 
     if (one->to_real() == two->to_real())
@@ -225,9 +225,9 @@ ALObjectPtr Fneq(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *ev
 {
     AL_CHECK(assert_min_size<0>(obj));
 
-    const auto one = eval_check(eval, obj, 0, &assert_number<int>);
+    const auto one = eval_check(eval, obj, 0, &assert_number<size_t>);
 
-    const auto two = eval_check(eval, obj, 1, &assert_number<int>);
+    const auto two = eval_check(eval, obj, 1, &assert_number<size_t>);
 
 
     if (one->to_real() != two->to_real())
@@ -245,9 +245,9 @@ ALObjectPtr Fmod(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *ev
 {
     AL_CHECK(assert_size<2>(obj));
 
-    const auto one = eval_check(eval, obj, 0, &assert_int<int>);
+    const auto one = eval_check(eval, obj, 0, &assert_int<size_t>);
 
-    const auto two = eval_check(eval, obj, 1, &assert_int<int>);
+    const auto two = eval_check(eval, obj, 1, &assert_int<size_t>);
 
     auto res = one->to_int() % two->to_int();
 
@@ -258,9 +258,9 @@ ALObjectPtr Fpow(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *ev
 {
     AL_CHECK(assert_size<2>(obj));
 
-    const auto one = eval_check(eval, obj, 0, &assert_number<int>);
+    const auto one = eval_check(eval, obj, 0, &assert_number<size_t>);
 
-    const auto two = eval_check(eval, obj, 1, &assert_number<int>);
+    const auto two = eval_check(eval, obj, 1, &assert_number<size_t>);
 
 
     return make_object(std::pow(one->to_real(), two->to_real()));
@@ -342,8 +342,8 @@ ALObjectPtr Fround(const ALObjectPtr &obj, env::Environment *, eval::Evaluator *
 {
     AL_CHECK(assert_size<2>(obj));
 
-    const auto one = eval_check(eval, obj, 0, &assert_number<int>);
-    const auto two = eval_check(eval, obj, 1, &assert_int<int>);
+    const auto one = eval_check(eval, obj, 0, &assert_number<size_t>);
+    const auto two = eval_check(eval, obj, 1, &assert_int<size_t>);
 
 
     return make_double(utility::round_nplaces(one->to_real(), two->to_int()));
