+++ src/alisp/src/definitions/alisp_stream_functions.cpp
@@ -44,7 +44,7 @@ ALObjectPtr Fstream(const ALObjectPtr &t_obj, env::Environment *, eval::Evaluato
 ALObjectPtr Fclose_stream(const ALObjectPtr &t_obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<1>(t_obj));
-    auto obj = eval_check(eval, t_obj, 0, &assert_stream<int>);
+    auto obj = eval_check(eval, t_obj, 0, &assert_stream<size_t>);
 
 
     StreamsHelper::close_stream(obj);
@@ -54,7 +54,7 @@ ALObjectPtr Fclose_stream(const ALObjectPtr &t_obj, env::Environment *, eval::Ev
 ALObjectPtr Fwith_cout(const ALObjectPtr &t_obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_min_size<1>(t_obj));
-    auto obj = eval_check(eval, t_obj, 0, &assert_stream<int>);
+    auto obj = eval_check(eval, t_obj, 0, &assert_stream<size_t>);
 
 
     StreamsHelper::rebind_cout(obj);
@@ -67,7 +67,7 @@ ALObjectPtr Fwith_cout(const ALObjectPtr &t_obj, env::Environment *, eval::Evalu
 ALObjectPtr Fwith_cin(const ALObjectPtr &t_obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_min_size<1>(t_obj));
-    auto obj = eval_check(eval, t_obj, 0, &assert_stream<int>);
+    auto obj = eval_check(eval, t_obj, 0, &assert_stream<size_t>);
 
     StreamsHelper::rebind_cin(obj);
 
@@ -79,7 +79,7 @@ ALObjectPtr Fwith_cin(const ALObjectPtr &t_obj, env::Environment *, eval::Evalua
 ALObjectPtr Fstream_content(const ALObjectPtr &t_obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<1>(t_obj));
-    auto obj    = eval_check(eval, t_obj, 0, &assert_stream<int>);
+    auto obj    = eval_check(eval, t_obj, 0, &assert_stream<size_t>);
     auto stream = StreamsHelper::get_stream(obj);
     return make_string(stream->content());
 }
@@ -87,7 +87,7 @@ ALObjectPtr Fstream_content(const ALObjectPtr &t_obj, env::Environment *, eval::
 ALObjectPtr Fstream_write(const ALObjectPtr &t_obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<2>(t_obj));
-    auto stream_obj = eval_check(eval, t_obj, 0, &assert_stream<int>);
+    auto stream_obj = eval_check(eval, t_obj, 0, &assert_stream<size_t>);
 
     auto str_obj = eval->eval(t_obj->i(1));
 
@@ -106,7 +106,7 @@ ALObjectPtr Fstream_write(const ALObjectPtr &t_obj, env::Environment *, eval::Ev
 ALObjectPtr Fstream_write_line(const ALObjectPtr &t_obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<2>(t_obj));
-    auto stream_obj = eval_check(eval, t_obj, 0, &assert_stream<int>);
+    auto stream_obj = eval_check(eval, t_obj, 0, &assert_stream<size_t>);
 
     auto str_obj = eval->eval(t_obj->i(1));
     auto stream  = StreamsHelper::get_stream(stream_obj);
@@ -124,7 +124,7 @@ ALObjectPtr Fstream_write_line(const ALObjectPtr &t_obj, env::Environment *, eva
 ALObjectPtr Fstream_write_lines(const ALObjectPtr &t_obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<2>(t_obj));
-    auto stream_obj = eval_check(eval, t_obj, 0, &assert_stream<int>);
+    auto stream_obj = eval_check(eval, t_obj, 0, &assert_stream<size_t>);
 
     auto stream = StreamsHelper::get_stream(stream_obj);
 
@@ -149,9 +149,9 @@ ALObjectPtr Fstream_write_lines(const ALObjectPtr &t_obj, env::Environment *, ev
 ALObjectPtr Fstream_read(const ALObjectPtr &t_obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<2>(t_obj));
-    auto stream_obj = eval_check(eval, t_obj, 0, &assert_stream<int>);
+    auto stream_obj = eval_check(eval, t_obj, 0, &assert_stream<size_t>);
 
-    auto int_obj = eval_check(eval, t_obj, 1, &assert_int<int>);
+    auto int_obj = eval_check(eval, t_obj, 1, &assert_int<size_t>);
 
 
     auto stream  = StreamsHelper::get_stream(stream_obj);
@@ -162,7 +162,7 @@ ALObjectPtr Fstream_read(const ALObjectPtr &t_obj, env::Environment *, eval::Eva
 ALObjectPtr Fstream_read_line(const ALObjectPtr &t_obj, env::Environment *, eval::Evaluator *eval)
 {
     AL_CHECK(assert_size<1>(t_obj));
-    auto stream_obj = eval_check(eval, t_obj, 0, &assert_stream<int>);
+    auto stream_obj = eval_check(eval, t_obj, 0, &assert_stream<size_t>);
 
     auto stream  = StreamsHelper::get_stream(stream_obj);
     auto content = (*stream).get_line();
@@ -173,7 +173,7 @@ ALObjectPtr Fstream_read_lines(const ALObjectPtr &t_obj, env::Environment *, eva
 {
 
     AL_CHECK(assert_size<1>(t_obj));
-    auto stream_obj = eval_check(eval, t_obj, 0, &assert_stream<int>);
+    auto stream_obj = eval_check(eval, t_obj, 0, &assert_stream<size_t>);
 
     auto stream = StreamsHelper::get_stream(stream_obj);
